modules/emoji/requirements.txt
requests
time
threaded
discord.py==1.7.0
Pillow
imgpy
opencv-python

